Rename checking._checking -> _name
CMK-12298
